requirements.txt
autoflake==2.3.1
beautifulsoup4==4.12.3
black==24.4.0
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
distlib==0.3.8
filelock==3.13.4
flake8==7.0.0
google-api-core==2.18.0
google-api-python-client==2.125.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
httplib2==0.22.0
identify==2.5.35
idna==3.7
isort==5.13.2
mccabe==0.7.0
mypy==1.9.0
mypy-extensions==1.0.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.0
pre-commit==3.7.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pyflakes==3.2.0
pyparsing==3.1.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
soupsieve==2.5
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.25.2
